backend/requirements.txt
fastapi==0.115.6
motor==3.6.0
python-dotenv==1.0.1
bcrypt==4.2.1
PyJWT==2.10.1
pydantic[email]==2.10.6
python-multipart==0.0.19
emergentintegrations
slowapi==0.1.9
passlib==1.7.4
email-validator==2.2.0